requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
requests>=2.31.0
resend>=0.1.0
passlib[argon2]>=1.7.4
python-jose>=3.3.0
argon2-cffi>=23.1.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
